pyproject.toml
[project]
name = "sfgraph"
version = "0.1.0"
description = "Salesforce Org Graph Analyzer — static analysis via property graph"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "sfgraph maintainers" },
]
keywords = ["salesforce", "mcp", "graph", "apex", "impact-analysis"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "qdrant-client==1.17.1",
    "fastembed==0.8.0",
    "aiosqlite==0.22.1",
    "pydantic>=2.0,<3.0",
    "mcp[cli]==1.27.0",
    "duckdb>=1.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<0.25",
    "pytest-cov>=5.0,<6.0",
]

[project.scripts]
sfgraph = "sfgraph.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sfgraph"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=sfgraph --cov-report=term-missing"
markers = [
    "integration: marks tests that require an external service (FalkorDB server on localhost:6379). Deselect with -m 'not integration'.",
]

[tool.coverage.run]
source = ["src/sfgraph"]
omit = ["tests/*"]
